Lots of message related changes. Removed WINDOWS ifdefs from messages.c. Changed message parser to use const char * strings. Changed base->deathtime to base->appeartime. Removed FPSDivisor variable. Fixed misc. warnings.

+++ xpilot/src/client/NT/wintalk.c
@@ -292,7 +292,7 @@ void Talk_delete_emphasized_text()
  *
  * Return the number of pasted characters.
  */
-int Talk_paste(char *data, int data_len, bool overwrite)
+int Talk_paste(char *data, size_t data_len, bool overwrite)
 {
 
     int str_len;		/* current length */
